+++ kratos/tests/test_KratosCore.py
@@ -41,7 +41,6 @@ import test_variable_component
 import test_variable_redistribution
 import test_object_printing
 import test_array_1d_interface
-import test_linear_master_slave_constraints
 import test_flags
 import test_time_discretization
 import test_python_generic_function_utility
@@ -50,7 +49,7 @@ import test_file_logger
 import test_dofs
 import test_time_averaging
 import test_scipy_conversion_tools
-import test_linear_constraints 
+import test_linear_constraints
 
 def AssembleTestSuites():
     ''' Populates the test suites to run.
@@ -108,7 +107,6 @@ def AssembleTestSuites():
     smallSuite.addTest(test_variable_redistribution.VariableRedistributionTest('testNodalArea'))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_object_printing.TestObjectPrinting]))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_array_1d_interface.TestArray1DInterface]))
-    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_linear_master_slave_constraints.TestLinearMultipointConstraints]))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_flags.TestFlags]))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_time_discretization.TestTimeDiscretization]))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_python_generic_function_utility.TestPythonGenericFunctionUtility]))
@@ -117,6 +115,7 @@ def AssembleTestSuites():
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_dofs.TestDofs]))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_time_averaging.TimeAveragingProcessTests]))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_scipy_conversion_tools.TestScipyConversionTools]))
+    smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_linear_constraints.TestLinearMultipointConstraints]))
     smallSuite.addTests(KratosUnittest.TestLoader().loadTestsFromTestCases([test_linear_constraints.TestLinearConstraints]))
 
     # Create a test suite with the selected tests plus all small tests
